+++ db/migrate/20251118025355_add_game_title_to_games.rb
@@ -1,0 +1,5 @@
+class AddGameTitleToGames < ActiveRecord::Migration[7.1]
+  def change
+    add_column :games, :game_title, :string
+  end
+end