+++ db/migrate/20251118033928_add_show_to_public_to_games.rb
@@ -1,0 +1,5 @@
+class AddShowToPublicToGames < ActiveRecord::Migration[7.1]
+  def change
+    add_column :games, :show_to_public, :boolean, default: false
+  end
+end